Adding support for decimal as data type.

+++ lib/Cake/Test/Case/Model/Datasource/Database/MysqlTest.php
@@ -545,7 +545,7 @@ class MysqlTest extends CakeTestCase {
 		$this->assertEquals($expected, $result);
 
 		$result = $this->Dbo->column('decimal(14,7) unsigned');
-		$expected = 'float';
+		$expected = 'decimal';
 		$this->assertEquals($expected, $result);
 	}
